Ajustes Parte 2

+++ src/ramos/monteiro/alex/service/CaixaServiceImpl.java
@@ -2,12 +2,15 @@ package ramos.monteiro.alex.service;
 
 import java.util.List;
 
+import ramos.monteiro.alex.model.Livro;
 import ramos.monteiro.alex.model.Produto;
+import ramos.monteiro.alex.model.Usuario;
+import ramos.monteiro.alex.utils.Utils;
 
 public class CaixaServiceImpl implements CaixaService {
-	
-	private static Double  caixa = 0.0;
-	
+
+	private static Double caixa = 0.0;
+
 	LojaServiceImpl lojaServiceImpl = new LojaServiceImpl();
 
 	@Override
@@ -15,10 +18,25 @@ public class CaixaServiceImpl implements CaixaService {
 		CaixaServiceImpl.caixa += valor;
 	}
 
+	//TODO: ajustar caso nao exista mais o produto. Por hora eu cobro e nao entrego Rs rs rs.
 	@Override
-	public void venderProduto(List<? extends Produto>listProduto, Produto produto) {
-		this.receberCaixa(produto.getPreco());
-		this.lojaServiceImpl.removeProduto(listProduto, produto);	
+	public void venderProduto(List<? extends Produto> listProduto, Produto produto, Usuario usuario) {
+
+		if (!produto.getIndicador()) {
+
+			this.calculaDesconto(produto);
+			this.receberCaixa(produto.getPreco());
+			this.lojaServiceImpl.removeProduto(listProduto, produto);
+
+		} else {
+			if (Utils.calculaIdade(usuario.getDataNascimento())) {
+				this.calculaDesconto(produto);
+				this.receberCaixa(produto.getPreco());
+				this.lojaServiceImpl.removeProduto(listProduto, produto);
+			}else {
+				System.out.println("Produto: " + produto.getNome() + " destinado a Maiores de 18 anos");
+			}
+		}
 	}
 
 	@Override
@@ -28,6 +46,13 @@ public class CaixaServiceImpl implements CaixaService {
 
 	public static Double getCaixa() {
 		return caixa;
+	}
+
+	private void calculaDesconto(Produto produto) {
+		if (produto.getClass().equals(Livro.class)) {
+			produto.setPreco(produto.getPreco() * 0.85);
+		}
+
 	}
 
 }
